Move PREPARE_FOR_ARKIT operator
