+++ human/expression/operators.py
@@ -3,10 +3,11 @@
 from pathlib import Path
 
 import bpy
-from HumGen3D.backend import hg_delete, remove_broken_drivers, get_prefs
+from HumGen3D.backend import get_prefs, hg_delete, remove_broken_drivers
 from HumGen3D.human.base.drivers import build_driver_dict
-
 from HumGen3D.human.human import Human
+from HumGen3D.human.human import Human
+from HumGen3D.user_interface.documentation.feedback_func import show_message
 
 
 class HG_REMOVE_SHAPEKEY(bpy.types.Operator):
@@ -89,4 +90,40 @@ class HG_REMOVE_FRIG(bpy.types.Operator):
     def execute(self, context):
         human = Human.from_existing(context.object)
         human.expression.remove_facial_rig()
+        return {"FINISHED"}
+
+
+class HG_OT_PREPARE_FOR_ARKIT(bpy.types.Operator):
+    bl_idname = "hg3d.prepare_for_arkit"
+    bl_label = "Prepare for ARKit"
+    bl_description = "Removes drivers and adds single keyframe to all FACS shapekeys"
+    bl_options = {"UNDO"}
+
+    suffix: bpy.props.EnumProperty(
+        name="Shapekey suffix",
+        items=[
+            ("long", "Left and Right (Default ARKit)", "", 0),
+            ("short", "_L and _R (FaceApp)", "", 1),
+        ],
+    )
+
+    def invoke(self, context, event):
+        return context.window_manager.invoke_props_dialog(self)
+
+    def draw(self, context):
+        layout = self.layout
+        layout.prop(self, "suffix", text="Suffix")
+
+    def execute(self, context):
+        human = Human.from_existing(context.object)
+
+        for sk in human.keys:
+            if sk.name == "Basis" or sk.name.startswith("cor_"):
+                continue
+            sk.driver_remove("value")
+            sk.keyframe_insert("value", frame=0)
+            if self.suffix == "long" and sk.name.endswith(("_L", "_R")):
+                sk.name = sk.name.replace("_L", "Left").replace("_R", "Right")
+
+        show_message(self, "Succesfully removed drivers and added keyframes")
         return {"FINISHED"}
